+++ frameworks/solid/src/types/field.ts
@@ -68,6 +68,10 @@ export interface FieldStore<
    */
   readonly isTouched: boolean;
   /**
+   * Whether the field value has been edited.
+   */
+  readonly isEdited: boolean;
+  /**
    * Whether the field input differs from its initial value.
    */
   readonly isDirty: boolean;
@@ -110,6 +114,10 @@ export interface FieldArrayStore<
    * Whether the field array has been touched.
    */
   readonly isTouched: boolean;
+  /**
+   * Whether the field array value has been edited.
+   */
+  readonly isEdited: boolean;
   /**
    * Whether the field array input differs from its initial value.
    */